refactor `createStorachaClient()`
