+++ bin/spark-evaluate.js
@@ -12,10 +12,7 @@ import pg from 'pg'
 import { createContracts } from '../lib/contracts.js'
 import { setScores } from '../lib/submit-scores.js'
 import { runPublishRsrLoop } from '../lib/publish-rsr.js'
-import * as Client from '@web3-storage/w3up-client'
-import { ed25519 } from '@ucanto/principal'
-import { CarReader } from '@ipld/car'
-import { importDAG } from '@ucanto/core/delegation'
+import { createStorachaClient } from '../lib/storacha.js'
 
 const {
   SENTRY_ENVIRONMENT = 'development',
@@ -38,21 +35,10 @@ assert(STORACHA_PROOF, 'STORACHA_PROOF required')
 
 await migrateWithPgConfig({ connectionString: DATABASE_URL })
 
-async function parseProof (data) {
-  const blocks = []
-  const reader = await CarReader.fromBytes(Buffer.from(data, 'base64'))
-  for await (const block of reader.blocks()) {
-    blocks.push(block)
-  }
-  return importDAG(blocks)
-}
-
-const principal = ed25519.Signer.parse(STORACHA_SECRET_KEY)
-const storachaClient = await Client.create({ principal })
-const proof = await parseProof(STORACHA_PROOF)
-const space = await storachaClient.addSpace(proof)
-await storachaClient.setCurrentSpace(space.did())
-
+const storachaClient = await createStorachaClient({
+  secretKey: STORACHA_SECRET_KEY,
+  proof: STORACHA_PROOF
+})
 const { ieContract, rsrContract, provider } = createContracts()
 
 const signer = ethers.Wallet.fromPhrase(WALLET_SEED, provider)
